refactor: extract image attachment logic into useImageAttachments hook
Extract image attachment state and handlers from MessageInput into a reusable hook. This improves code organization and makes the attachment logic available for other components (e.g., NewTaskBar).
